+++ src/components/navBar/navBar.js
@@ -17,6 +17,7 @@ import gitHubLogo from './../../assets/githubLogo/github.png';
 import viadeLogo from './../../assets/logo/logo_alt.jpeg';
 import viadeText from './../../assets/logo/logo_letters.jpeg';
 import RouteManager from "./../../model/RouteManager";
+import ShareView from '../../pages/ShareView';
 
 const routeManager = new RouteManager();
 
@@ -78,7 +79,7 @@ function MyNavBar(props) {
       <Route exact path="/friends" component={Friends} />
       <Route exact path="/editProfile" component={EditProfile} />
       <Route exact path="/routes/info/:id" render={(props) => <InfoView routeManager={routeManager} {...props} />} />
-      <Route exact path="/routes/share/:id" render={(props) => <InfoView routeManager={routeManager} {...props} />} />
+      <Route exact path="/routes/share/:id" render={(props) => <ShareView routeManager={routeManager} {...props} />} />
       <Redirect path="/" exact to="/home" />
     </HashRouter>
   );
